requirements.txt
antspyx==0.6.1
einops==0.6.1
fastremap==1.14.1
hiddenlayer==0.3
loguru==0.7.2
connected-components-3d==3.16.0
matplotlib==3.5.3
MedPy==0.4.0
natsort==8.4.0
nibabel==4.0.2
numpy==1.21.6
pandas==1.3.5
batchgenerators==0.25
Pillow==9.5.0
prefetch_generator==1.0.3
pytorch_metric_learning==2.3.0
#PyYAML==6.0.1
scikit_learn==1.0.2
scipy==1.8.0
SimpleITK==1.2.4
timm==0.6.13
wandb==0.18.3
tqdm==4.65.0
yacs==0.1.8
fvcore==0.1.5.post20221221
torchinfo==1.8.0